requirements.txt
discord-ext-ipc==2.1.1
discord.py==2.4.0
DateTime==5.4
PyNaCl==1.5.0